+++ tests/testthat/test-visualize-enrichment-chart.R
@@ -78,6 +78,18 @@ test_that("`enrichment_chart()` -- produces a ggplot object with correct labels"
   expect_equal(labels$colour, expression(-log[10](p)))
   expect_equal(labels$x, "Fold Enrichment")
   expect_equal(labels$y, "Term_Description")
+
+  # Change order_by
+  expect_is(g <- enrichment_chart(example_pathfindR_output_clustered, order_by = "highest_p"),
+      "ggplot")
+  expect_equal(ggplot2::quo_name(g$mapping$x), "Fold_Enrichment")
+  expect_equal(ggplot2::quo_name(g$mapping$y), "Term_Description")
+  
+  labels <- ggplot2::get_labs(g)
+  expect_equal(labels$size, "# genes")
+  expect_equal(labels$colour, expression(-log[10](p)))
+  expect_equal(labels$x, "Fold Enrichment")
+  expect_equal(labels$y, "Term_Description")
 })
 
 test_that("`enrichment_chart()` -- argument checks work", {
@@ -105,5 +117,6 @@ test_that("`enrichment_chart()` -- argument checks work", {
     "`top_terms` must be either numeric or NULL"
   )
 
+  expect_error(enrichment_chart(example_pathfindR_output, order_by = "INVALID"))
   expect_error(enrichment_chart(example_pathfindR_output, top_terms = 0), "`top_terms` must be > 1")
 })